user account styled

+++ Client/src/components/Profile/Account.js
@@ -1,21 +1,23 @@
-import React, { useContext, useEffect } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import Header from "../Header.js";
 import { MyContext } from "../../MyContext.js";
 import { useNavigate } from "react-router-dom";
-import { NavLink, useLocation } from "react-router-dom";
-import bookNook from "../../assets/bookNook.jpg"
+import { useLocation } from "react-router-dom";
+import bookNook from "../../assets/bookNook.jpg";
+import AvatarDropdown from "./AvatarDropdown.js";
 
 function Account() {
     const { user, loading, error, setError, deleteAccount } = useContext(MyContext);
-    const { bookclubs, bookshelves } = user || {};
+    const { bookclubs, bookshelves, accolades, goals, streak, avatar } = user || {};
+
     const navigate = useNavigate();
-    const location = useLocation()
+    const location = useLocation();
 
     useEffect(() => {
-        if (!user) {
-            navigate('/login')
+        if (!user && !loading) {
+            navigate('/login');
         }
-    });
+    }, [user, loading, navigate]);
 
     useEffect(() => {
         setError(null); 
@@ -23,9 +25,9 @@ function Account() {
 
     const backgroundStyle = {
         backgroundImage: `url(${bookNook})`,
-        backgroundSize: "100%",
+        backgroundSize: "cover", 
         backgroundPosition: "center",
-        minHeight: "100vh",
+        minHeight: "120vh",
         filter: "blur(6px)",
         opacity: 0.6,
         backgroundRepeat: "no-repeat",
@@ -35,11 +37,13 @@ function Account() {
         right: 0,
         bottom: 0,
         zIndex: -1,
-      };
-    
+    };
+
     const handleDeleteAccount = () => {
-        deleteAccount(user.id)
-    }
+        if (window.confirm("Are you sure you want to delete your account?")) {
+            deleteAccount(user.id);
+        }
+    };
 
     return (
         <div>
@@ -50,22 +54,17 @@ function Account() {
             {/* Background div */}
             <div style={backgroundStyle}></div>
 
-            {/* Loading and Error States */}
-            {loading && <p className="text-center text-xl">Loading...</p>}
-            {error && <p className="text-center text-red-500 text-xl">{error}</p>}
-
             {/* Full-screen container */}
-            <div className="hero  min-h-screen flex flex-col items-center justify-center">
-                {/* Profile content container */}
-                <div className="hero-content flex flex-col items-center space-y-6 w-full max-w-4xl px-4">
-                    {/* Welcome message */}
+            <div className="hero min-h-screen flex flex-col items-center justify-center">
+                <div className="hero-content flex flex-col items-center space-y-6 w-full max-w-4xl px-4 relative">
+                    {/* Welcome Message */}
                     <h1 className="text-5xl font-bold text-center">
                         Welcome Back {user?.username}
                     </h1>
 
-                    {/* Avatar section */}
-                    <div className="relative">
-                        {user?.avatar ? (
+                    {/* Avatar Section */}
+                    <div className="relative flex flex-col items-center space-y-4">
+                        {user && user.avatar ? (
                             <img
                                 src={user.avatar}
                                 alt="User Avatar"
@@ -74,90 +73,96 @@ function Account() {
                         ) : (
                             <div className="avatar placeholder">
                                 <div className="bg-primary text-neutral-content w-44 rounded-full">
-                                    <span className="text-2xl">
-                                        {user?.username ? user.username : 'UN'}
-                                    </span>
+                                    <span className="text-2xl">{user?.username ? user.username : 'UN'}</span>
                                 </div>
                             </div>
                         )}
+                    </div>
 
-                        {/* Edit Button */}
-                        <div className="absolute bottom-2 left-1/2 transform -translate-x-1/2 flex gap-2">
-                            <NavLink to="/avatar">
-                                <button className="btn btn-secondary btn-sm">Edit</button>
-                            </NavLink>
-                        </div>
+                    {/* Edit Avatar Button */}
+                    <div className=""> {/* Adjust top and right values as needed */}
+                        <AvatarDropdown currentAvatar={user?.avatar} />
                     </div>
 
                     {/* Main Content */}
                     <div className="space-y-6 w-full">
-                        {/* Bookshelves Section */}
-                        <div className="flex justify-between items-center w-full">
-                            <h1 className="text-3xl font-bold">Bookshelves</h1>
-                            <NavLink to={"/bookshelves"}>
-                            <button className="btn btn-primary btn-sm">Go To Your Bookshelves</button>
-                            </NavLink>
-                        </div>
-
-                        {/* Bookshelves Carousel */}
-                        <div className="border p-4 rounded-lg bg-white shadow-lg">
-                            <div className="flex overflow-x-auto space-x-4">
-                                {bookshelves && bookshelves.length > 0 ? (
-                                    bookshelves.map((shelf, index) => (
-                                        <div key={index} className="w-60 p-4 bg-white rounded-xl shadow-lg flex-none">
-                                            <h2 className="text-xl font-semibold">{shelf.name}</h2>
-                                            <p>{shelf.description}</p>
-                                            <button className="btn btn-primary btn-sm mt-4">View Shelf</button>
-                                        </div>
-                                    ))
-                                ) : (
-                                    <div className="col-span-3 text-center">
-                                        <p>Currently no books on your shelves</p>
-                                    </div>
-                                )}
+                        {/* Accolades Carousel */}
+                        <div className="space-y-4">
+                            <h2 className="text-3xl font-bold text-primary">Accolades</h2>
+                            <div className="flex items-center space-x-4">
+                                <div className="overflow-x-auto space-x-4">
+                                    {accolades && accolades.length > 0 ? (
+                                        accolades.map((accolade, index) => (
+                                            <div
+                                                key={index}
+                                                className="w-60 p-4 bg-neutral text-white rounded-xl shadow-lg flex-none"
+                                            >
+                                                <h3 className="text-xl font-semibold">{accolade.title}</h3>
+                                                <p>{accolade.description}</p>
+                                            </div>
+                                        ))
+                                    ) : (
+                                        <p>No accolades yet.</p>
+                                    )}
+                                </div>
                             </div>
                         </div>
 
-                        {/* Bookclubs Section */}
-                        <div className="flex justify-between items-center w-full">
-                            <h1 className="text-3xl font-bold">Bookclubs</h1>
-                            <NavLink to={"/users/:userId/bookclubs"}>
-                            <button className="btn btn-primary btn-sm">Go To Your Bookclubs</button>
-                            </NavLink>
+                        {/* Goals Carousel */}
+                        <div className="space-y-4">
+                            <h2 className="text-3xl font-bold text-primary">Goals</h2>
+                            <div className="flex items-center space-x-4">
+                                <div className="overflow-x-auto space-x-4">
+                                    {goals && goals.length > 0 ? (
+                                        goals.map((goal, index) => (
+                                            <div
+                                                key={index}
+                                                className="w-60 p-4 bg-neutral text-white rounded-xl shadow-lg flex-none"
+                                            >
+                                                <h3 className="text-xl font-semibold">{goal.name}</h3>
+                                                <p>{goal.description}</p>
+                                            </div>
+                                        ))
+                                    ) : (
+                                        <p>No goals set yet.</p>
+                                    )}
+                                </div>
+                                <div className="ml-4">
+                                    {/*</div>button className="btn btn-secondary btn-sm">Edit Goals</button> */}
+                                </div>
+                            </div>
                         </div>
 
-                        {/* Bookclubs Carousel */}
-                        <div className="border p-4 rounded-lg bg-white shadow-lg">
-                            <div className="flex overflow-x-auto space-x-4">
-                                {bookclubs && bookclubs.length > 0 ? (
-                                    bookclubs.map((club, index) => (
-                                        <div key={index} className="w-60 p-4 bg-white rounded-xl shadow-lg flex-none">
-                                            <h2 className="text-xl font-semibold">{club.name}</h2>
-                                            <p>{club.description}</p>
-                                            <button className="btn btn-primary btn-sm mt-4">Join Club</button>
-                                        </div>
-                                    ))
+                        {/* Streak Section */}
+                        <div className="space-y-4">
+                            <h2 className="text-3xl font-bold text-primary">Reading Streak</h2>
+                            <p className="text-xl">
+                                {streak > 0 ? (
+                                    <>Reading Streak: <span className="font-bold text-primary">{streak}</span> days!</>
                                 ) : (
-                                    <div className="col-span-3 text-center">
-                                        <p>No Bookclubs found, please join a bookclub.</p>
-                                    </div>
+                                    <span className="text-white">No reading streaks yet.</span>
                                 )}
-                            </div>
+                            </p>
                         </div>
 
                         {/* Member Since */}
-                        <h1 className="text-xl font-bold">Member since: {new Date(user?.created_at).toLocaleDateString('en-US', { year: 'numeric', month: 'long', day: 'numeric' })}</h1>
+                        <h1 className="text-xl font-bold">
+                            Member since: {new Date(user?.created_at).toLocaleDateString('en-US', { year: 'numeric', month: 'long', day: 'numeric' })}
+                        </h1>
                     </div>
-                    <button onClick={handleDeleteAccount} className="btn btn-primary btn-sm">Delete Account</button>
+
+                    {/* Delete Account Button */}
+                    <button onClick={handleDeleteAccount} className="btn btn-danger btn-secondary btn-sm">Delete Account</button>
                 </div>
             </div>
-        {/* Footer */}
-      <footer className="bg-white py-6 border-t-4 text-black">
-        <div className="container mx-auto text-center">
-          <p>&copy; 2025 The Tattered Page. All rights reserved. Made with ❤️ for book lovers</p>
+
+            {/* Footer */}
+            <footer className="bg-white py-6 border-t-4 text-black mt-auto"> {/* mt-auto will push the footer to the bottom */}
+                <div className="container mx-auto text-center">
+                    <p>&copy; 2025 The Tattered Page. All rights reserved. Made with ❤️ for book lovers</p>
+                </div>
+            </footer>
         </div>
-      </footer>
-    </div>
     );
 }
